+++ src/main/java/com/imsweb/algorithms/prcda/PrcdaDataProvider.java
@@ -30,25 +30,7 @@ import static com.imsweb.algorithms.prcda.PrcdaUtils.PRCDA_YES;
  */
 public class PrcdaDataProvider {
 
-    private enum PrcdaYear {
-        YR2017, YR2020;
-
-        public static PrcdaYear getDefault() {
-            return YR2020;
-        }
-    }
-
     public String getPrcda(String state, String county) {
-        return getPrcdaData(PrcdaYear.getDefault(), state, county);
-    }
-
-    public String getPrcda2017(String state, String county) {
-        return getPrcdaData(PrcdaYear.YR2017, state, county);
-    }
-
-    private String getPrcdaData(PrcdaYear prcdaYear, String state, String county) {
-        prcdaYear = prcdaYear == null ? PrcdaYear.getDefault() : prcdaYear;
-
         if (ENTIRE_STATE_PRCDA.contains(state))
             return PRCDA_YES;
         else if (ENTIRE_STATE_NON_PRCDA.contains(state))
@@ -67,10 +49,7 @@ public class PrcdaDataProvider {
         if (countyData == null)
             return PRCDA_NO;
 
-        if (PrcdaYear.YR2017.equals(prcdaYear))
-            return countyData.getPrcda2017();
-        else
-            return countyData.getPrcda();
+        return countyData.getPrcda();
     }
 
     private Map<String, Map<String, CountyData>> loadPrcdaData() {
@@ -80,11 +59,9 @@ public class PrcdaDataProvider {
 
             String state = line.getField(0);
             String county = line.getField(1);
-            String prcda2017 = line.getField(2);
-            String prcda2020 = line.getField(3);
+            String prcda2024 = line.getField(2);
             CountyData dto = result.computeIfAbsent(state, k -> new HashMap<>()).computeIfAbsent(county, k -> new CountyData());
-            dto.setPrcda2017(StringUtils.leftPad(prcda2017, 1, '0'));
-            dto.setPrcda(StringUtils.leftPad(prcda2020, 1, '0'));
+            dto.setPrcda(StringUtils.leftPad(prcda2024, 1, '0'));
         });
 
         return result;