+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 **Changes in version 5.11 (not released yet)**
 
 - Added 2023 calculated field for Rural Urban Continuum algorithm.
+- Updated the lookup for item #194 in the PRCDA algorithm to use 2024 PRCDA counties
+- Removed the 2017 calculated field from the PRCDA algorithm.
 - Added missing census2020 input field to Tract-Estimated Congressional Districts; this only affects documentation, the logic was actually using the field.
 - Updated dependencies.